Transferred the alert logic to the new view controller, replaced it with the logic that displays the alert on the timeline. Buttons aren't receiving events, for some reason.
